DRF implementation/requirements.txt
absl-py==2.1.0
aio-pika==8.3.0
aiofiles==22.1.0
aioredis==2.0.1
aiormq==6.6.4
aiosqlite==0.17.0
alembic==1.13.1
amqp==5.2.0
ansible==7.2.0
ansible-core==2.14.14
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.14.0
bcrypt==4.0.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Django==4.2.11
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-admin==1.0.4
fastapi-cache2==0.2.1
fastapi-pagination==0.12.19
fastapi-utils==0.2.1
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
influxdb-client==1.36.0
iso8601==1.1.0
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
joblib==1.3.2
jwcrypto==1.5.6
kiwisolver==1.4.5
kombu==5.3.5
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.11
minio==7.1.13
ml-dtypes==0.3.2
motor==3.3.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pamqp==3.2.1
pandas==2.2.1
passlib==1.7.4
pendulum==3.0.0
pika==1.3.1
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psycopg2-binary==2.9.6
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.4
pyfcm==1.5.4
PyJWT==2.8.0
pymongo==4.6.2
pyparsing==3.1.2
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2024.1
PyYAML==6.0.1
rarfile==4.1
reactivex==4.0.4
redis==5.0.3
requests==2.31.0
resolvelib==0.8.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.42.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
SQLAlchemy==1.4.52
SQLAlchemy-Utils==0.41.1
sqlparse==0.5.0
starlette==0.32.0.post1
sympy==1.12
threadpoolctl==3.3.0
time-machine==2.14.0
torch==2.2.2
tortoise-orm==0.20.0
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.20.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1
